Rename iOS benchmark device runner
