+++ AppBenchDeviceRunner/AppBenchDeviceRunner/AppBenchDeviceRunnerApp.swift
@@ -1,6 +1,6 @@
 //
-//  FoundationStudioApp.swift
-//  FoundationStudio
+//  AppBenchDeviceRunnerApp.swift
+//  AppBenchDeviceRunner
 //
 //  Created by Rudrank Riyam on 10/25/25.
 //
@@ -8,7 +8,7 @@
 import SwiftUI
 
 @main
-struct FoundationStudioApp: App {
+struct AppBenchDeviceRunnerApp: App {
     var body: some Scene {
         WindowGroup {
             ContentView()
