+++ AppBenchDeviceRunner/AppBenchDeviceRunner/Assets.xcassets/AccentColor.colorset/Contents.json
