+++ AppBenchDeviceRunner/AppBenchDeviceRunner/Assets.xcassets/AppIcon.appiconset/Contents.json
