+++ AppBenchDeviceRunner/AppBenchDeviceRunner/Assets.xcassets/Contents.json
